Update CID to v1.3
Some sites were coming back as false positives as a result of the UA, so
it's been replaced with a UA from the most up to date Chrome browser.
Some sites, despite the fact they should be able to handle them, only
exclusively support the + or the %20 in the query string to represent a
whitespace.  The most common usage is %20 which is what the sanitizer
sanitizes to, but some sites exlcusively use the + instead in which case
it would break the URL and deliver a bad result.  So the option to
preserve the existing encoding format has been added to the sanitizer
and full requests now preserve the encoding.
Also fixed an encoding parameter.  Could be a possible issue, that we
should avoid.